feat(calculate-debt) auto debt capacity calculation if loan have automatic validation

+++ domain/model/src/main/java/co/com/bancolombia/model/application/gateways/DebtCapacityGateway.java
@@ -5,5 +5,5 @@ import reactor.core.publisher.Mono;
 import java.util.UUID;
 
 public interface DebtCapacityGateway {
-    Mono<String> loanDecision(double totalIncome, double currentMonthlyDebt, double interestRate, int termMonths, float loanAmount, UUID idApplication, String email);
+    Mono<String> loanDecision(Long totalIncome, double currentMonthlyDebt, double interestRate, int termMonths, float loanAmount, UUID idApplication,String email);
 }
